requirements.txt
pytest
chromedriver
selenium
allure-pytest
